Add the Benchmarks Games benchmarks
These were modified to be run without exiting Erlang and to return
output in a more convenient way to check. This definitely modifies the
benchmarks as accommodations needed to be made to remove the
`erlang:halt/1` calls and hardcoded writes. However, they do the
original calculation, so they're close.

The versions that require input are commented out.